model_manager_lib/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
clang==5.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
dataclasses-json==0.5.4
flatbuffers==1.12
gast==0.4.0
google-api-core==2.0.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-cloud-core==2.0.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
grpcio==1.39.0
grpcio-tools==1.39.0
h5py==3.1.0
idna==3.2
iniconfig==1.1.1
json-logging==1.3.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
marshmallow==3.13.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pathy==0.6.0
pluggy==0.13.1
pprintpp==0.4.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-clarity==1.0.1
pytest-instafail==0.4.2
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.9.0
rsa==4.7.2
six==1.15.0
smart-open==5.2.0
stringcase==1.2.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-serving-api==2.5.1
termcolor==1.1.0
toml==0.10.2
typer==0.3.2
typing-extensions==3.7.4.3
typing-inspect==0.7.1
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
